Send buttons like a regular message
So apparently, the theory is you need to add components to The Message Builder, and then send a message, but as a button. Untested to work, but this should send buttons like a normal message, just with different functionality than a normal message.

Also, in order for us to do this, we needed to implement channel ID to actually fetches the Discord channel by the Channel ID, and then using that to send the message.