Added full support for Discord Interactions!
This commit now fully supports Discord interactions. Please note - It should be stable enough to download, but may contain bugs. If you ever receive a bug with addons, then please create an issue. Thank you!

To those that want to read this -
* You can now modify Interactions. The differences between Modifying Interactions and sending a message, is with Interactions, you're only editing the Message Builder, which ofcourse doesn't have the ID built-in, which is what you use sendMessage() for.
SendMessage, only includes The Message class objective. The Message Builder is not currently built-in to sendMessage() packet, as apart of our backwards-compatibility. We may change this in the future, but for now, this can do.

* We've added Interaction Models back because they're apart of an "Parts" which extends the Part class.
** Also, for those that didn't realize, we've added Option Model with Interaction models as that's apart of a part, and should now be complete. Fire your code! Let's see what custom things you can do with them! :)

Also keep note - For anyone wondering, I'll update the API or make examples on how to properly use the Interaction API soon.