Fixed Bot not being compatible for other bots.
This commit, is a fix for Certain bots from not being compatible with other bots.
For example, Tebex, sometimes your bot would turn off if you have Tebex in your server. Basically, anything that has the "url" in the discord bots that weren't either twitch or youtube, this would then become a problem. Because we can't really identify if whether or not the url is custom, we'll justify that for you. Basically, this commit adds more support than only twitch or youtube.