Deprecated stickers in favour of sticker_items
This means any plugins that use DiscordMessage->stickers will now need to use sticker_items due to discord's API being changed. This issue had a bug, where Stickers would be null. This was because again, it was deprecated in favour of stickers, so all should be working as intended. This update has been tested to work.