Added Member Timeout semi-support
And I say semi, because I need to implement this directly onto the plugin. It is possible to implement this outside our Discord API, but currently, this is not possible via inside the core API.

This is a semi-support due to that. There is no estimate time until this get's supported fully by the API.

- Alongside this, add new Moderate members permissions node.