Use MessageBuilder + Message->edit() to edit the message, instead of using safe part method
This commit fully supports Stickers, TTS, and Objective mentions (Channels, roles, and users) to Editing a message.

I know I said it was kinda pointless, but this needed a redo anyways, as this was using an outdated API source, which was built before Message Builder even existed. Eventually, we'll need to require Message Builder objectives. But for now, Message Builder will be inside the Editing a message method, instead of as a parameter.