Permissions + Bitwise Update (Untested)
This commit hugely improves on bitwise system, and how the permissions are calculated.

- Change permissions to only numbers, instead of displaying the duplication of numbers form.

- Require ext-gmp php extension. (If using 32 bit system)

- Added missing API information.

- Updated Readme: Now display a new requirement if using a 32 bit system.

- Check if using 32 bit system, and the extension ext-gmp is loaded. If not, it'll return an trigger_error.

- Added Bitwise Helper tool to Discord php.

- Rename permission from use_slash_commands -> use_application_commands

- Rename permission from use_public_thread -> create_public_thread.

- Rename permission from use_private_thread -> create_private_thread.

- bitwise should also return string if using 32 bit.

- Add Attribute functions for Permission class.

- Member: Use Bitwise import.

- Member: Use Permission import.

- Permissions Model: Throw an error if bitwise is below 0.

- ModelConverter: Updated to support the new bitwise system.

**NOTE**
This is a complete permissions rewrite. In the next update, it'll be a brand new major version, which changes permissions as a whole.

Also to note, this plugin release version, it will most probably be buggy, and untested. If you appear to find any issues with this update, please do open an issue! It will be released shortly.