Fixed Storage::isOldMessage() incorrect return values
This fix was really needed, as the isOldMessage() function would return with incorrect return types. This commit basically fixes this issue.

Also, the reason why I made this change, is because Storage->getMessagesByChannel() only fetches all messages since the last 14 days, which means once the foreach() part is out of the way, it'll return true, instead of false - As it is considered old if messages aren't foreached with the message object you want, as discord doesn't store messages older than 14 days (2 weeks).

- Added Storage::removeOldMessage()
* Automatically deletes messages that are older than 14 days.
* We add this because the bot will only check for messages older than 14 days upon start up, which would've caused possible errors in the future.
* Now, it'll automatically detect messages regardless. It's on a task every 1 tick, which means it's always checking.