Fixed "This interaction failed" when interacting wthout weird code hacks
This is probably the most needed fix. In the future, I plan to deprecate a lot of Request interactions. You can use RequestRespondInteraction to respond to an interaction (This fixes The interaction failed error), RequestCreateInteraction (For creating an interaction AKA Sending a message builder message), and RequestModifyInteraction (For Modifying an interaction AKA Editing a message builder message.

The rest of the interactions will be deprecated soon.

- Added missing component type parameter, instead of making it as type (AKA One parameter, instead of two).

- Added RequestRespondInteraction.

- InteractionData Model: Fixed getComponentType() always being null.

- Resolved: Added typehints to array functions.

- CommandMessage: Fixed components issue.

- Added Api->createInteractionResponse().