Major BC: Re-organize namespaces + Clean-up a ton of folders, Classes/Models, and more
This commit is unfortunately a Breaking Change. I was going to do this once we support 4.0 API, but I decided to do this now.

This commit hugely improves on namespaces, modeling areas, adding Server\ Model namespaces, \User model namespaces, and etc. This is attempting to match Discord Part's name spacing whilst still making a couple of things in our own namespacing models.

- So I'll say it again - Re-located all Model namespaces to their appropriate form.
- The next change log will be marked as Breaking Change.

- It'll also detail what namespaces have been changed from and to.